// packages/shared/src/types.ts
export interface Feed {
    id: number;
    podcastGuid: string;
    title: string;
    url: string;
    originalUrl: string;
    link: string;
    description: string;
    author: string;
    ownerName: string;
    image: string;
    artwork: string;
    lastUpdateTime: number;
    lastCrawlTime: number;
    lastParseTime: number;
    lastGoodHttpStatusTime: number;
    lastHttpStatus: number;
    contentType: string;
    itunesId: number | null;
    itunesType: string;
    generator: string;
    language: string;
    explicit: boolean;
    type: 0 | 1;
    medium: string;
    dead: number;
    chash: string;
    episodeCount: number;
    crawlErrors: number;
    parseErrors: number;
    categories: Record<string, string>;
    locked: 0 | 1;
    imageUrlHash: number;
}

export interface Episode {
    id: number;
    title: string;
    link: string;
    description: string;
    guid: string;
    datePublished: number;
    datePublishedPretty: string;
    dateCrawled: number;
    enclosureUrl: string;
    enclosureType: string;
    enclosureLength: number;
    duration: number | null;
    explicit: 0 | 1;
    episode: number | null;
    episodeType: 'full' | 'trailer' | 'bonus' | null;
    season: number | null;
    image: string;
    feedItunesId: number | null;
    feedItunesType: string;
    feedUrl: string;
    feedImage: string;
    feedId: number;
    podcastGuid: string;
    feedLanguage: string;
    feedDead: number;
    feedDuplicateOf: number | null;
    chaptersUrl: string | null;
    transcriptUrl: string | null;
}